Move "Local Update Available" case to its own file
